feat(event): add layout component for event details with metadata generation
